Some refactor and cleanup
